Animations Updated
Updated cloud movement animation for the home menu scene. Also I will try to use more accurate terminology such as *scene* instead of **screen** and things to that effect so that these updates are more easily clear in what was done.

Working on new buttons since the text is hard to see on top of the button due to similar color choice for both vectors. New `play_button` added and will be style for placeholder menu items until graphics are ready.