Custom Transition Scene Added
Added a custom made scene transition node that handles the loading bar animation and display of the logo. The control of the flow is handled by a series of functions in a Global.gd script which checks first for any current scene activity, waits for it to end, and then appends the requested scene as a child of the current.