Menu Update, NewGameScene Update, and more
Lots of new updates to the setup for the game. Main menu cloud graphics replaced with actual clouds made for the project. Also separated out a scene for a new game and all moves made *after* the first one. Still need to build the next part which is path-finding and implement turn based navigation for the player. Grid is mostly setup by limiting players movement to 16x16 coordinate spaces in any direction. This should help with loading the proper number of "tiles" in  the "grid" to allow for proper spacing of players to the top of the screen.

Still need to add the sun and audio.