Sunlight Spot Generation & Other Updates
Still working on the turn based movement tests and need to build out the system for checking the compatibility between two sprites when building a branch. This is tricky but not impossible. 

Got sunlight to show on screen at start of game but still working on getting the size right.