Corrected Angle Calculations
Adjusted the angle calculation functions to take into account that Godot uses a more dynamic formula for calculating radians on the fly. The engine actually draws a 'ghost' triangle to complete the shape of a triangle that has the two Vectors at specified locations. 

Just had to subtract 90 from the degree value after having the `abs()` global function called on that.